+++ pkg/constants/meta.go
@@ -1,9 +1,9 @@
 package constants
 
-// MetaPrefix is the MetaPrefix for labels, annotations, and finalizers of Accurate.
+// MetaPrefix is the MetaPrefix for labels, annotations, and finalizers of Cattage.
 const MetaPrefix = "cattage.cybozu.io/"
 
-// Finalizer is the finalizer ID of Accurate.
+// Finalizer is the finalizer ID of Cattage.
 const Finalizer = MetaPrefix + "finalizer"
 
 const OwnerTenant = MetaPrefix + "tenant"
@@ -11,7 +11,3 @@ const OwnerTenant = MetaPrefix + "tenant"
 const OwnerAppNamespace = MetaPrefix + "owner-namespace"
 
 const TenantFieldManager = MetaPrefix + "tenant-controller"
-const ApplicationFieldManager = MetaPrefix + "application-controller"
-const StatusFieldManager = ApplicationFieldManager + "/status"
-const SpecFieldManager = ApplicationFieldManager + "/spec"
-const ProjectFieldManager = ApplicationFieldManager + "/project"